+++ src/app/pages/challenge/challenge.component.ts
@@ -5,7 +5,9 @@ import { catchError, switchMap } from 'rxjs/operators';
 import { PageTitleService } from '@sage-bionetworks/sage-angular';
 import {
   ChallengeService,
+  ChallengePlatformService,
   Challenge,
+  ChallengePlatform,
   ModelError as RoccClientError,
 } from '@sage-bionetworks/rocc-client-angular';
 import { isRoccClientError } from '@shared/rocc-client-error';
@@ -19,6 +21,7 @@ import { ChallengeDataService } from './challenge-data.service';
 export class ChallengeComponent implements OnInit {
   @HostBinding('class.main-content') readonly mainContentClass = true;
   challenge$!: Observable<Challenge | undefined>;
+  platform$!: Observable<ChallengePlatform | undefined>;
   challengeNotFound = false;
 
   sections = [
@@ -36,6 +39,7 @@ export class ChallengeComponent implements OnInit {
     private router: Router,
     private route: ActivatedRoute,
     private challengeService: ChallengeService,
+    private challengePlatformService: ChallengePlatformService,
     private pageTitleService: PageTitleService,
     private challengeDataService: ChallengeDataService
   ) {}
@@ -61,6 +65,9 @@ export class ChallengeComponent implements OnInit {
       this.pageTitleService.setTitle(`${pageTitle} · ROCC`);
       this.challengeNotFound = !challenge;
       this.challengeDataService.setChallenge(challenge);
+      // this.platform$ = this.challengePlatformService.getChallengePlatform(
+      //   challenge.platformId
+      // );
     });
   }
 }
